Modify RRE server to use service to supply evaluation data.
