+++ rre-server/src/main/java/io/sease/rre/server/RREServer.java
@@ -2,6 +2,8 @@ package io.sease.rre.server;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.core.env.ConfigurableEnvironment;
+import org.springframework.core.env.StandardEnvironment;
 
 /**
  * RRE Server main entry point.
@@ -12,6 +14,10 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class RREServer {
     public static void main(final String[] args) {
-        SpringApplication.run(RREServer.class, args);
+        SpringApplication application = new SpringApplication(RREServer.class);
+        ConfigurableEnvironment environment = new StandardEnvironment();
+        environment.setDefaultProfiles("http");
+        application.setEnvironment(environment);
+        application.run(args);
     }
 }
