+++ rre-server/src/main/resources/static/modules/main/config-service.js
@@ -15,7 +15,7 @@
          * The data request URL
          * @type {string}
          */
-        var requestUrl = "http://127.0.0.1:8080/evaluation";
+        var requestUrl = "/evaluation";
 
         init();
